Sync Arcadia yt/python -> GitHub

+++ python-git/yt/wrapper/job_runner.py
@@ -93,7 +93,13 @@ def make_run_script(destination_dir):
         path = path[:-1]
 
     destination_path = os.path.join(destination_dir, "run")
-    shutil.copy2(path, destination_path)
+    if os.path.exists(path):
+        shutil.copy2(path, destination_path)
+    else:
+        import library.python.resource
+        with open(destination_path, "w") as f:
+            f.write(library.python.resource.find("/job_runner_py_source"))
+
     os.chmod(destination_path, os.stat(destination_path).st_mode | stat.S_IXUSR)
 
 if __name__ == "__main__":
